backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.0
python-dotenv==1.0.0